third_party/blink/web_tests/scrollbars/scrollevent-iframe-no-scrolling-wheel.html

<!DOCTYPE html>
<script src="../resources/gesture-util.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
var receivedScrollEvent = false;
var t = async_test('The iframe cannot be scolled by WheelEvent.');

function scrollEventFired() {
    receivedScrollEvent = true;
}

async function runTest() {
    window.frames[0].onscroll = scrollEventFired;
    var doc = window.frames[0].document;
    var fillDiv = doc.getElementById("fillDIV");
    var center = elementCenter(fillDiv);
    await wheelTick(0, 1, center);
    await conditionHolds( () => {
        return window.frames[0].scrollY == 0;
    });
    assert_false(receivedScrollEvent, "no scroll event is received");
    t.done();
}
</script>

<iframe onload="runTest();" scrolling="no" style="width:200px;height:200px;" src="resources/scrollable-iframe.html">
